Moved Action Log controller to Backend namespace
