+++ app/Http/Controllers/Backend/ActionLogController.php
@@ -2,8 +2,9 @@
 
 declare(strict_types=1);
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Backend;
 
+use App\Http\Controllers\Controller;
 use App\Services\ActionLogService;
 
 class ActionLogController extends Controller
